agent-system/backend/requirements.txt
django==4.2 
djangorestframework==3.14 
django-cors-headers==4.3 
python-dotenv==1.0 
requests==2.31 
anthropic==0.18 
toml==0.10 
